fix: create /terms route, fix company name to Recase Inc, redirect /privacy to /terms
- Create /terms route showing Terms of Service content
- Rename privacyContent.ts to termsContent.ts for accuracy
- Fix all references from 'Rebase, Inc.' to 'Recase, Inc.'
- Redirect /privacy to /terms (no privacy policy content exists)
- Update agentContent.ts to use termsContent and buildTermsMarkdown
- Add /terms.md route for markdown content negotiation
