+++ apps/website/lib/termsContent.ts
@@ -1,16 +1,16 @@
-export type PrivacyTerm = {
+export type TermsItem = {
   title: string;
   content: string;
   isUppercase: boolean;
 };
 
-export const PRIVACY_EFFECTIVE_DATE = "February 1, 2025";
+export const TERMS_EFFECTIVE_DATE = "February 1, 2025";
 
-export const privacyTerms: PrivacyTerm[] = [
+export const termsOfService: TermsItem[] = [
   {
     title: "1. Acceptance of Terms",
     content:
-      'By accessing or using the Autumn platform, APIs, SDKs, or any related services (collectively, the "Services") provided by Rebase, Inc., a Delaware corporation ("Autumn," "we," "us," or "our"), you agree to be bound by these Terms of Service ("Terms"). If you do not agree to these Terms, do not use our Services.',
+      'By accessing or using the Autumn platform, APIs, SDKs, or any related services (collectively, the "Services") provided by Recase, Inc., a Delaware corporation ("Autumn," "we," "us," or "our"), you agree to be bound by these Terms of Service ("Terms"). If you do not agree to these Terms, do not use our Services.',
     isUppercase: false,
   },
   {
@@ -100,7 +100,7 @@ export const privacyTerms: PrivacyTerm[] = [
   {
     title: "16. Contact Information",
     content:
-      "For questions about these Terms, please contact us at security@useautumn.com or at: Rebase, Inc. (d/b/a Autumn), Email: security@useautumn.com, Website: https://useautumn.com",
+      "For questions about these Terms, please contact us at security@useautumn.com or at: Recase, Inc. (d/b/a Autumn), Email: security@useautumn.com, Website: https://useautumn.com",
     isUppercase: false,
   },
 ];
